Maven plugin configuration changes for generating the app Change the individual artifacts to be under org.springframework.cloud Change stream -> dataflow in artifact id's
