+++ spring-cloud-starter-dataflow-collector-metrics/src/main/java/org/springframework/cloud/dataflow/metrics/collector/MetricCollectorProperties.java
@@ -1,0 +1,35 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.dataflow.metrics.collector;
+
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+/**
+ * @author Vinicius Carvalho
+ */
+@ConfigurationProperties(prefix = "spring.cloud.dataflow.metrics.collector")
+public class MetricCollectorProperties {
+	private Integer evictionTimeout = 30;
+
+	public Integer getEvictionTimeout() {
+		return evictionTimeout;
+	}
+
+	public void setEvictionTimeout(Integer evictionTimeout) {
+		this.evictionTimeout = evictionTimeout;
+	}
+}
