+++ spring-cloud-starter-dataflow-collector-metrics/src/main/java/org/springframework/cloud/dataflow/metrics/collector/utils/YANUtils.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2017 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.springframework.cloud.dataflow.metrics.collector.utils;
+
+import java.util.Optional;
+
+/**
+ * Yet Another Utils class
+ * @author Vinicius Carvalho
+ */
+public class YANUtils {
+	public static <S, T> Optional<T> cast(S candidate, Class<T> targetClass) {
+		return targetClass.isInstance(candidate)
+				? Optional.of(targetClass.cast(candidate))
+				: Optional.empty();
+	}
+}
