+++ spring-cloud-starter-dataflow-collector-metrics/src/test/java/org/springframework/cloud/dataflow/metrics/collector/CollectorApplicationTests.java
@@ -1,0 +1,16 @@
+package org.springframework.cloud.dataflow.metrics.collector;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+@RunWith(SpringRunner.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class CollectorApplicationTests {
+
+	@Test
+	public void contextLoads() {
+	}
+
+}
