--- spring-cloud-starter-stream-collector-metrics/src/main/java/org/springframework/cloud/stream/app/metrics/collector/model/Application.java
@@ -1,75 +1,0 @@
-/*
- * Copyright 2017 the original author or authors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.springframework.cloud.stream.app.metrics.collector.model;
-
-import java.util.LinkedList;
-import java.util.List;
-
-import com.fasterxml.jackson.annotation.JsonCreator;
-
-/**
- * @author Vinicius Carvalho
- */
-public class Application {
-	private String name;
-
-	@JsonCreator
-	public Application(String name) {
-		this.name = name;
-	}
-
-	private List<Instance> instances = new LinkedList<>();
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public List<Instance> getInstances() {
-		return instances;
-	}
-
-	public void setInstances(List<Instance> instances) {
-		this.instances = instances;
-	}
-
-	public Double getIncomingRate(){
-		return getInstances().stream().mapToDouble(Instance::getIncomingRate).sum();
-	}
-
-	public Double getOutgoingRate(){
-		return getInstances().stream().mapToDouble(Instance::getOutgoingRate).sum();
-	}
-
-	@Override
-	public boolean equals(Object o) {
-		if (this == o) return true;
-		if (o == null || getClass() != o.getClass()) return false;
-
-		Application that = (Application) o;
-
-		return name != null ? name.equals(that.name) : that.name == null;
-	}
-
-	@Override
-	public int hashCode() {
-		return name != null ? name.hashCode() : 0;
-	}
-}
