--- spring-cloud-starter-stream-collector-metrics/src/test/java/org/springframework/cloud/stream/app/metrics/collector/CollectorApplicationTests.java
@@ -1,16 +1,0 @@
-package org.springframework.cloud.stream.app.metrics.collector;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class CollectorApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
